+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - Decrease default timeout to `500 ms`
+- Improve error message formatting
 
 ## [0.10.0] - 2026-02-09